move IntervalValue out of types subdir
